A messy commit that implements basic testing
This commit is not clean. It was a very quick attempt to get some
numbers and should be used with some caution. Some of the highlights of
the changes include:
- Add nodefile for Luigi
- Update `experiment.proto` to use a `WorkloadProto` and include
connection information like the public and private hostnames for each
node in the cluster
- Update `client.h`, `server.h` and `main.cc` to use new protos
- Hacked together `launch.py` to launch experiments (should be reviewed
and cleaned up) and `plot.py` to plot the results
- Changed `setup.h` to define the save path of results

NOTE: During this process I discovered that there was something odd
going on with Rome's `ConnectionManager` with respect to loopback. I'll
file a bug separately.